:white_check_mark: Add tests
* :white_check_mark: Test for module `convert_between_variables`
* :white_check_mark: Test for module `synthetic_volcanoes`
* :white_check_mark: Test for module `view_force`
* :sparkles: Better CLI argument for `view-frc`, check for file
  existence and accepts no argument -> default to last created.

+++ tests/test_convert_between_variables.py
@@ -8,7 +8,7 @@ from volcano_cooking.modules import convert
 def test_vei_to_totalemission() -> None:
     a = np.arange(100)
     b = [1, 2, 3, 4]
-    c = np.arange(10, dtype=np.float32)
+    c = np.arange(10, dtype=np.int8)
     with pytest.raises(ValueError):
         convert.vei_to_totalemission(a)
         convert.vei_to_totalemission(b)
